pyproject.toml
[project]
name = "vibe-photos"
version = "1.0.0"
description = "AI-powered photo management system for content creators"
requires-python = "==3.12.*"
license = "AGPL-3.0-or-later"
authors = [{ name = "jasl" }]

dependencies = [
    "celery>=5.5.3",
    "sqlalchemy>=2.0.44",
    "typer>=0.20.0",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "aiofiles>=25.1.0",
    "pydantic>=2.12.4",
    "requests>=2.32.5",
    "pyyaml>=6.0.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.1",
    "sentence-transformers>=5.1.2",
    "datasets>=4.4.1",
    "numpy>=2.3.5",
    "accelerate>=1.11.0",
    "flask>=3.1.2",
    "xxhash>=3.6.0",
    "inflect>=7.0.0",
    "openai>=1.0.0",
]

[tool.uv]
package = true

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "httpx>=0.28.1",
    "black>=25.11.0",
    "ruff>=0.14.5",
    "mypy>=1.18.2",
    "isort>=7.0.0",
    "ipython>=9.7.0",
    "ipdb>=0.13.13",
    "rich>=14.2.0",
    "python-dotenv>=1.2.1",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.0",
    "pdoc>=16.0.0",
    "memory-profiler>=0.61.0",
    "line-profiler>=5.0.0",
    "py-spy>=0.4.1",
]
phase_final = [
    "fastapi>=0.121.2",
    "uvicorn>=0.38.0",
    "streamlit>=1.51.0",
    "psycopg2-binary>=2.9.11",
    "pgvector>=0.4.1",
    "redis>=7.0.1",
    "prometheus-client>=0.23.1",
]

[tool.uv.sources]
# Install PyTorch with CUDA support on Linux/Windows (CUDA doesn't exist for Mac).
# NOTE: We must explicitly request them as `dependencies` above. These improved
# versions will not be selected if they're only third-party dependencies.
torch = [
  { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cuda"
# Use PyTorch built for NVIDIA Toolkit version 13.0.
# Available versions: https://pytorch.org/get-started/locally/
url = "https://download.pytorch.org/whl/cu130"
# Only use this index when explicitly requested by `tool.uv.sources`.
explicit = true

[tool.ruff]
target-version = "py312"
line-length = 150

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # do not perform function calls in argument defaults
]

[tool.black]
line-length = 150
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "8.4.2"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py"]
